fix(playground): moved models.ts to seperate directory
