+++ playground/components/ExampleComponent.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { computed, ref } from 'vue';
-import { Todo, Meta } from './models';
+import { Todo, Meta } from '../shared/models';
 
 interface Props {
   title: string;
